feat(website): fold the live Tier B climb into the landing hero
The hero terminal now does both jobs with one chrome: the recorded climb
plays on load, and a "Try it on a real site" CTA swaps the same terminal
to a live Tier B run (HTTP to a headed Chromium via a residential proxy)
against a curated bot-protected target, gated by email + Turnstile.

- New HeroClimb owns the single ClimbDemo and the reveal: the canned
  script until a live run is submitted, then the live SSE stream.
- Drop the standalone /playground page and its now-orphaned components
  (PlaygroundShowcase, the Tier A LiveClimb, LiveClimbTierB). Nothing
  linked to it (not in nav or sitemap; it was noindex).
- climb-demo: unreached rungs read "not needed" once the climb concludes,
  instead of a lingering "queued" (only a live browser-rung win hits this;
  the canned demos reach their top rung).

The live affordance stays behind NEXT_PUBLIC_PLAYGROUND_TIERB_LIVE, so the
hero is a pure recording until the flag is flipped.

Verified: pnpm build + tsc clean; canned hero + reveal + email gate; a real
end-to-end Rozetka climb through the gate (HTTP 403, browser 200, origin
Markdown), about 47s warm.
